Move api & stringify to their own file
